+++ action_server/docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add package build argument --input-dir to choose the action package source directory
 - Package build command to print package path to stdout in json format if --json argument is provided
 - Encrypt-at-rest sensitive storage data to the file system instead of using keyring to store it